+++ astroquery/gaia/tests/setup_package.py
@@ -10,7 +10,8 @@ def get_package_data():
     paths = [os.path.join('data', '*.vot'),
              os.path.join('data', '*.vot.gz'),
              os.path.join('data', '*.json'),
-             os.path.join('data', '*.ecsv')
+             os.path.join('data', '*.ecsv'),
+             os.path.join('data', '*.zip')
              ]  # etc, add other extensions
     # you can also enlist files individually by names
     # finally construct and return a dict for the sub module